fix: correct terminal workflow evidence integration
